Move device implementations
